Demo on G3MApp

+++ iOS/G3MiOSSDK/Commons/Rendererers/PlanetRenderer.cpp
@@ -488,7 +488,7 @@ RenderState PlanetRenderer::getRenderState(const G3MRenderContext* rc) {
   if (!_allFirstLevelTilesAreTextureSolved) {
     _firstLevelTilesJustCreated = false;
 
-    const int firstLevelTilesCount = _firstLevelTiles.size();
+    const size_t firstLevelTilesCount = _firstLevelTiles.size();
 
     if (_tilesRenderParameters->_forceFirstLevelTilesRenderOnStart) {
       _statistics.clear();
